bug fix, improve gameplay a little bit, add code to check player has won or not
fixed the bug, which caused the player to shoot continuously on pressing down SPACE, even if I check just for change in events, not the current event(it may be a bug of SDL2.0). I have reduced the max number of enemy tanks in level 1, to make the first level easier.